Circle shaded cells in PuzzleMe format.
If there are no explicitly circled cells, but there are cells with
different background colors, then circle all cells with different
background colors. Also ensure we don't circle both isCircled cells as
well as cells with custom background shapes if both exist in the same
puzzle, since the former is a higher-fidelity signal that those cells
should be circled, and combining the two could be confusing.