Fix 10/6/2019 Washington Post puzzle.
This puzzle had hidden letters in some black squares which are
revealed upon entering the complete solution. The squares have letters
according to the grid field, but no words intersect them according to
the boxToPlacedWordsIdxs field. (They are also set to have black
foreground text until the puzzle is completed, which effectively makes
the fill invisible, but this is a subtler effect to try to parse).