Add Javascript support for PDF generation.
Factor out the PDF assembly logic from the format logic, and add
expect/actual implementations using PDFBox (as before) for Java and
jsPDF for Javascript. For unit testing, we use PDF.js to render the
expected/actual PDFs as images for comparison.

The golden PDF is nearly identical visually to the one from before -
indeed, the Java rendering logic sees no difference, but the Javascript
rendering logic sees a slight difference, so we update the golden to use
the Javascript version instead.