Add support for UTF-8 Across Lite files.
These files use version 2.0 and decode metadata and clues as UTF-8.

When writing as Across Lite, we only use the 2.0 version (and UTF-8
encoding) if the puzzle contains characters that cannot be encoded in
ISO-8859-1 and if the caller doesn't force use of ISO-8859-1. If the
puzzle contains unencodable characters and the caller forces use of
ISO-8859-1, the puzzle text is fully sanitized.