Escape unsupported HTML tags.
PuzzleMe JSON has markup HTML (e.g. <i> tags) escaped exactly the same
way as non-markup HTML which is meant to be rendered literally (e.g. the
clue "<div> or </div>, in HTML", from today's Atlantic puzzle. So, when
we escape PuzzleMe clues but leave HTML tags in tact, we only allow HTML
tags that are part of the JPZ spec, since we assume those are the only
tags which will be used for formatting.