Migrate to IR compiler.
Reduces the size of the compiled Javascript in ~half. PDF generation
seems a fair bit slower, but otherwise everything seems to be in working
order. Requires that any methods accessed from JS are marked as
@JsExport and are public, so we have to remove the "internal" keyword
from the form classes. Instead, we add an opt-in annotation
@KotwordsExperimental and remove the classes from the docs to discourage
direct usage.

Supporting the legacy mode as well would require splitting the web
components into their own module, which is more complicated than it
seems worth if nobody's using it.

Fixes #8