Migrate from ktor-io to okio.
ktor-io is not particularly meant for external use, and the primary
benefit of using it was the ISO-8859-1 support which is buggy and not
actually correct for decoding, where we want Cp-1252.

okio is actively maintained and meant for public use. The main drawback
is that it only supports UTF-8, so we need to implement Cp-1252
encoding/decoding ourselves, which is fairly straightforward. We can
also replace our manual Base64 implementation since okio supports Base64
encoding/decoding.

When decoding, we use Cp-1252, which adds support for a few characters
that were previously unknown. When encoding, we also use it, but we
consider any Cp-1252-specific characters to require UTF-8, since these
are only supported on Windows Across Lite, not Mac. For now, if UTF-8 is
forced off, these characters will be encoded in Cp-1252. As part of
adding general fallbacks for UTF-8, we should add replacements for the
remaining Cp-1252 characters to be cross platform safe.

Fixes #20