Use unidecode data to generalize mapping of Unicode to ISO-8859-1.
We generate a 256x256 array in Kotlin code using a Perl script that
checks the Unicode transliteration from unidecode for each character
code. The transliteration itself is a trivial application of the mapping
to each character.

Unlike unidecode, we leave characters in the 0x80-0xFF range untouched;
while these are not valid ASCII, they are valid ISO-8859-1 and are thus
supported by Across Lite. We also apply a few mappings from our previous
hard-coded map which were missing from the base data.

Note that these mappings are slightly different - for example, Greek
letters are transliterated into their English equivalents, rather than
translated into their literal names - but the output is overall as
reasonable as it can probably be, and easier to maintain.

Fixes #2