Add support for importing Coded crosswords from .puz files.
Adds an "Import PUZ" button to the bottom of the form. When a .puz file
is imported, the title, creator, copyright, description, and grid are
set to those of the .puz file. Further edits can then be made directly
in the form (and saved as a JSON file for future use, as usual).

See #21