Fix regression in Acrostic completion message propagation.
In c086256, the APZ file form started taking the completion message from
the manual entry form by mistake. Swap it back to the APZ file form as
intended. Note that as of this refactoring, the message is no longer
provided to createPuzzleFn, but is needed when determining whether to
generate a completion message or leave it blank, since the Puzzle's
completion message takes precedence over the applet settings when
generating JPZ files. So we expose a getter for the completion message.