PuzzleMe Marching Bands support.
There doesn't seem to be a canonical indicator in the PuzzleMe data
itself; instead, an extra snippet of Javascript is included in the HTML
for these puzzles, presumably controlled server-side. As an initial
heuristic, we detect Marching Bands puzzles based on a regex matching of
the clues, but this could be tweaked in the future as needed.

Also includes fixes for newlines in PuzzleMe descriptions and support
for multi-line descriptions in PDFs.

Fixes #34