Handle word lengths and anchor tags in PuzzleMe.
Anchor tags are replaced with just their text contents, since they can't
be shown properly in a valid JPZ file.

Word lengths are mapped to the JPZ format attribute if enabled at a file
level, and rendered directly into the clue when exported to PUZ/PDF. As
a consequence, we now support the format attribute in JPZ. However, a
number of sources in the wild tend to set format unnecessarily or
inconsistently, so simply propagating it from these sources may not be
ideal. This is ultimately a decision to be made at a per-source level,
so we expose a JPZ parsing option to strip the format attributes.

See jpd236/CrosswordScraper#26