Revamp puzzle conversion interfaces.
- Move common public format conversion API methods into Puzzleable. This
allows file format readers (Across Lite and JPZ) to support directly
returning the input data without modification (or risk of parse/output
failure) when the requested output is in the same format. It also allows
converters to directly output into a format that has native support for
a particular puzzle type without having to force it into the Puzzle
format, which may be useful in the future e.g. for iPuz Acrostics.

- Add DelegatingPuzzleable helper for parsers which work by fitting data
into another model class. For example, this allows Acrostic parsers to
directly return Acrostic and take advantage of future optimizations like
native iPuz or PDF support, instead of returning a converted Puzzle. It
also allows parsers to support more than one type of puzzle, such as
PuzzleMe's support for Rows Garden puzzles in addition to the standard
Puzzle support.

- Add JpzFile class to more cleanly encapsulate an actual JPZ file,
since the existing JPZ class is a representation of the parsed data.
This handles the unzip and XML parsing as needed, and also allows us to
represent the raw data in a way that can be returned directly when
converting JPZ to JPZ, without depending on parsing.