Initial implementation of Ipuz support.
Supports read and write for most variety puzzles. Coded crosswords do
not appear to be supported directly. Acrostics are supported, but the
format documentation is too ambiguous to implement, so this may come
later.

HTML forms have been extended to support Ipuz generation. The "compress
JPZ" option has been renamed to "Compress File", though for Ipuz, this
only controls whether or not we enable pretty printing, since Ipuz files
are not intended to be compressed.