Migrate from jspdf to pdf-lib.
The unfortunate driving factor for this is that jspdf contains a
reference to an external script in its save method, which is disallowed
in Chrome extensions, even if the codepath is unused. Resolving this
would require forking jspdf, or having the logic moved out of the main
save method, neither of which feels great.

A pleasant side effect is that the generated PDFs are significantly
smaller, perhaps because jspdf doesn't subset the fonts. A drawback is
that the library size is significantly higher.

We also upgrade the Kotlin version since the prior one was encountering
compiler errors.