Add support for Patchwork puzzles.
Includes the "hard" mode which sorts the Pieces clues and leaves them
unassociated with their locations in the grid, and marks the ipuz as
having "fake clues" (a proprietary attribute defined by squares.io, the
only supported applet for such puzzles). Use a heuristic to control when
"fake clues" is set that appears to work for existing tests, but might
need further refinement to avoid falsely triggering it for quirky ipuz
puzzles.

Fixes #18