Support puzzles with no solution.
Across Lite has an explicit bit indicating when a solution is missing.
We now respect this bit and ignore the solution grid in this case. IPUZ
has an explicit "solution" field which can be omitted. JPZ does not have
an explicit way of omitting the solution, but Crossword Compiler omits
the check/reveal actions from the metadata and omits all cell values.

In Puzzle form, we heuristically determine that there is no solution by
seeing if every editable cell has an empty solution. A few empty cells
can be valid, but at least one should have a non-empty solution in a
valid puzzle.

Handle WSJ JSON without embedded solutions.