Support reading WSJ bylines from JSON "meta" field.
Acrostics no longer contain the byline in the previous location; it has
moved to meta.byline. Regular puzzles still have the byline in the usual
location, but have a second copy present in meta.author. In both cases,
prefer the entry in meta, but fall back to the prior field for older
data if the new entry is missing/null/empty.

Fixes jpd236/CrosswordScraper#38