Add support for Hearts and Arrows puzzles with unlabeled hearts.
This should also be usable for Rows Garden puzzles if we pass along the
appropriate setting to the common layer.

Also, disable annotation generation by default for Hearts and Arrows
since this is a somewhat peculiar quirk of Rows Gardens, and move
heart/bloom labels to the top right to put them in an actual corner and
move them a bit further from the row labels for clarity.