Initial implementation of Kotlin/Native support.
Includes all features except PDF generation and puzzle cell images.
Builds a placeholder binary for Linux, Windows, and macOS (all X64). In
the future, this binary will be a command line utility exposing useful
Kotwords features for dumping data, conversions, etc.

The GitHub CI build should now run on Mac/Windows, solely for the native
targets of those platforms. JS/JVM builds continue to be exclusive to
Linux for CI purposes.

Implementation notes:

- Zip files use korio. This is technically a multiplatform solution, but
newer versions of the library were merged into a much larger game engine
package, Korge, and are not currently available as a standalone package.
Perhaps we can use this across platforms if this is resolved.

- HTML/XML parsing uses ksoup. This is a multiplatform solution which is
effectively a drop-in replacement for jsoup. It may be suitable for all
platforms as it matures, if the size/speed overhead on JS/JVM is
minimal.

- ksoup currently depends on kotlinx-datetime 0.5.0 which seems to cause
our build to fail if we depend on 0.6.0-RC.2 (which is useful for
replacing klock). For now we can continue to hold off.

- URL encoding uses a small standalone library which could be usable on
all platforms for simplicity.