Support HTML constructor notes in Crosshare (from Slate).
For simplicity, rather than try and introduce HTML support, just extract
and concatenate all of the text fields together. If we ever see
legitimate HTML, we can try and leverage our handling of it for
JPZ/IPUZ, but at the moment it's riskier as this would mean treating the
clues as HTML as well.