Support rebuses and special characters in Marching Bands.
If a grid row contains no whitespace, we parse it as before, treating
each character as a cell. If it contains any whitespace (after
trimming), we split by whitespace. Special characters are now left
unchanged (though "." continues to be treated as a black square) rather
than being filtered out.

See #44