Fully support ipuz clue list titles.
In ipuz, a cluelist consists of a direction and an optional user-visible
string. For example, "Diagonal:Southeast" tells an application that the
clues in the list go down and to the right per the "Diagonal", but that
it should display the string "Southeast" as the user-visible title.

We handle this by treating the user-visible title as the title in the
existing Puzzle data class, and adding a new "direction" field to cover
the first element of a two-part clue list title.

See #47