Handle ipuz files with 0-based coordinates.
When reading ipuz files, use heuristics to detect whether 0-based
coordinates are being used, and handle them correctly if so. In case of
any doubt or ambiguity, we continue to use 1-based coordinates as
before.

We still write 1-based coordinates. If/when a new version of the ipuz
spec clarifies that 0-based coordinates are to be used, we will need to
support this tied to the requested output file version.

See #46