Implement native PDF and image support.
Replaces the platform-specific PDF and image handling via helper
libraries with a direct Kotlin PDF generator. We only need
a platform-specific helper for deflate. Images are handled by
korlibs-image-core, which is functional although seems to be
unmaintained.

Since the platform-specific implementations still create identical
outputs, we can remove the platform-specific test files in favor of
exact equivalence checks.

There are some minor API changes around moving everything to a .pdf
subpackage and tweaking the TTF specifications to allow for lazy loading
and separate the font identifiers from the font data.