Bump PuzzleMe max key digit to 19 and add back optional JS parsing.
If the crossword JS is provided, we use it to determine the set of
possible key digits. This is a small optimization, but my hope is it
would smooth things over temporarily if the max key digit changes again
rather than requiring an emergency patch. Significantly increasing the
maxKeyDigit slows down the brute forcing substantially, so we would need
a new approach altogether if that happens.