Use more clue lists in Two-Tone and Jelly Roll puzzles.
Since JPZ only allows up to two clue lists, we were combining the
light/dark squares lists into one, which is unnecessarily constrained
for other output formats (specifically IPUZ and PDF). We can provide the
full lists of clues along with a merge key that can be used to
consolidate the lists to two when needed, but keep them split apart
otherwise.

To start, this is just done for these two puzzle types, but there are
others (e.g. the Light/Medium/Dark Blooms in Rows Gardens) we could
expand this to if there are no issues discovered here.

See #52